Move chats direct/:id as routeParams
